# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  github_com_musashimiyomoto_todo-app_internal_core_transport_http_types.Nullable-bool:
    properties:
      set:
        type: boolean
      value:
        type: boolean
    type: object
  github_com_musashimiyomoto_todo-app_internal_core_transport_http_types.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
  internal_features_statistics_transport_http.GetStatisticsResponse:
    properties:
      tasksAverageCompletionTime:
        type: string
      tasksCompleted:
        type: integer
      tasksCompletedRate:
        format: float64
        type: number
      tasksCreated:
        type: integer
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        type: integer
      description:
        maxLength: 1000
        minLength: 1
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_types.Nullable-bool'
      description:
        $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_types.Nullable-string'
      title:
        $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_types.Nullable-string'
    type: object
  internal_features_tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_types.Nullable-string'
      phone_number:
        $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_types.Nullable-string'
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST API
  title: Todo App API
  version: "1.0"
paths:
  /statistics:
    get:
      description: Returns task statistics with optional filtering by user and date
        range.
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Statistics
          schema:
            $ref: '#/definitions/internal_features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get statistics
      tags:
      - Statistics
  /tasks:
    get:
      description: Returns a list of tasks with optional filtering and pagination.
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Creates a new task in the system.
      parameters:
      - description: Create task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Create task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      description: Deletes a task by ID.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Task deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Delete task
      tags:
      - Tasks
    get:
      description: Returns a single task by ID.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task found
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get task by ID
      tags:
      - Tasks
    patch:
      consumes:
      - application/json
      description: Partially updates a task by ID.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patch task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated successfully
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Patch task
      tags:
      - Tasks
  /users:
    get:
      description: Returns a list of users with optional pagination.
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user in the system.
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Deletes a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: User deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Delete user
      tags:
      - Users
    get:
      description: Returns a single user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Partially updates a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patch user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_musashimiyomoto_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Patch user
      tags:
      - Users
swagger: "2.0"