# internal/features/tasks/transport/http/delete_task.go
package tasks_transport_http

import (
	"net/http"

	core_logger "github.com/musashimiyomoto/todo-app/internal/core/core_logger"
	core_http_request "github.com/musashimiyomoto/todo-app/internal/core/transport/http/request"
	core_http_response "github.com/musashimiyomoto/todo-app/internal/core/transport/http/response"
)

// DeleteTask 	 godoc
// @Summary 	 Delete task
// @Description  Deletes a task by ID.
// @Tags 		 Tasks
// @Param 		 id path int true "Task ID"
// @Success 	 204 "Task deleted successfully"
// @Failure 	 400 {object} core_http_response.ErrorResponse "Bad request"
// @Failure 	 404 {object} core_http_response.ErrorResponse "Task not found"
// @Failure 	 500 {object} core_http_response.ErrorResponse "Internal server error"
// @Router 		 /tasks/{id} [delete]
func (h *TasksHTTPHandler) DeleteTask(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	taskID, err := core_http_request.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(err, "Failed to get 'id' path value")

		return
	}

	if err := h.tasksService.DeleteTask(ctx, taskID); err != nil {
		responseHandler.ErrorResponse(err, "Failed to delete task")

		return
	}

	responseHandler.NoContentResponse()
}