# internal/features/users/transport/http/get_users.go
package users_transport_http

import (
	"fmt"
	"net/http"

	core_logger "github.com/musashimiyomoto/todo-app/internal/core/core_logger"
	core_http_request "github.com/musashimiyomoto/todo-app/internal/core/transport/http/request"
	core_http_response "github.com/musashimiyomoto/todo-app/internal/core/transport/http/response"
)

type GetUsersResponse []UserDTOResponse

// GetUsers 	 godoc
// @Summary 	 Get users
// @Description  Returns a list of users with optional pagination.
// @Tags 		 Users
// @Produce 	 json
// @Param 		 limit  query int false "Limit"
// @Param 		 offset query int false "Offset"
// @Success 	 200 {object} GetUsersResponse "List of users"
// @Failure 	 400 {object} core_http_response.ErrorResponse "Bad request"
// @Failure 	 500 {object} core_http_response.ErrorResponse "Internal server error"
// @Router 		 /users [get]
func (h *UsersHTTPHandler) GetUsers(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	limit, offset, err := getLimitOffsetQueryParams(r)
	if err != nil {
		responseHandler.ErrorResponse(err, "Failed to get 'limit' and 'offset' query params")

		return
	}

	userDomains, err := h.usersService.GetUsers(ctx, limit, offset)
	if err != nil {
		responseHandler.ErrorResponse(err, "Failed to get users")

		return
	}

	responseHandler.JSONResponse(
		GetUsersResponse(
			usersDTOFromDomains(userDomains),
		),
		http.StatusOK,
	)
}

func getLimitOffsetQueryParams(r *http.Request) (*int, *int, error) {
	const (
		limitQueryParamKey  = "limit"
		offsetQueryParamKey = "offset"
	)

	limit, err := core_http_request.GetIntQueryParam(r, limitQueryParamKey)
	if err != nil {
		return nil, nil, fmt.Errorf("Failed to get 'limit' query param: %w", err)
	}

	offset, err := core_http_request.GetIntQueryParam(r, offsetQueryParamKey)
	if err != nil {
		return nil, nil, fmt.Errorf("Failed to get 'offset' query param: %w", err)
	}

	return limit, offset, nil
}